openswe cli: updates
